cleaned up css and busy icons
Issue around date filtering of events; there is no server-side ways to specify > and < with dates, so this implies client-side management, which (realistically) requires a new service to manage this data and the controller to talk to it instead. this may also help if we add a 'postcodes in view' service, which may be messier due to the postcode API not wanting to give us all the postcodes for the UK at once, which could happen in we zoom out enough.